+++ src/service/worker/runtime/graphql/schema/schemaDirectives/index.ts
@@ -1,10 +1,10 @@
 import { Auth, authDirectiveTypeDefs } from './Auth'
 import { CacheControl, cacheControlDirectiveTypeDefs } from './CacheControl'
 import { Deprecated, deprecatedDirectiveTypeDefs } from './Deprecated'
+import { Metric, metricDirectiveTypeDefs } from './Metric'
 import { SanitizeDirective, sanitizeDirectiveTypeDefs } from './Sanitize'
 import { SettingsDirective, settingsDirectiveTypeDefs } from './Settings'
 import { SmartCacheDirective, smartCacheDirectiveTypeDefs } from './SmartCacheDirective'
-import { Telemetry, telemetryDirectiveTypeDefs } from './Telemetry'
 import { TranslatableV2, translatableV2DirectiveTypeDefs } from './TranslatableV2'
 import { TranslateTo, translateToDirectiveTypeDefs } from './TranslateTo'
 
@@ -14,10 +14,10 @@ export const nativeSchemaDirectives = {
   auth: Auth,
   cacheControl: CacheControl,
   deprecated: Deprecated,
+  metric: Metric,
   sanitize: SanitizeDirective,
   settings: SettingsDirective,
   smartcache: SmartCacheDirective,
-  telemetry: Telemetry,
   translatableV2: TranslatableV2,
   translateTo: TranslateTo,
 }
@@ -26,10 +26,10 @@ export const nativeSchemaDirectivesTypeDefs = [
   authDirectiveTypeDefs,
   cacheControlDirectiveTypeDefs,
   deprecatedDirectiveTypeDefs,
+  metricDirectiveTypeDefs,
   sanitizeDirectiveTypeDefs,
   settingsDirectiveTypeDefs,
   smartCacheDirectiveTypeDefs,
-  telemetryDirectiveTypeDefs,
   translatableV2DirectiveTypeDefs,
   translateToDirectiveTypeDefs,
 ].join('\n\n')