+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `@telemetry` GraphQL directive
+- `@metric` GraphQL directive
 
 ## [6.43.1] - 2021-07-12